Create Korean.properties
I used a translator. 
Naturally, the quality of the translation is very poor.
But since there was no Korean translation, I think it has value right now.

It was translated quickly using a translator before starting the game, so there may be more mistranslations.
I plan to play the game and someday refine the content. But sure, This game is "civ" and it may not be a proper translation...

However, I believe it has been translated so that it can be read, so please register it officially.
I hope someone can do a proper translation...